Remove some compilation warning under msvc
git-svn-id: svn://svn.code.sf.net/p/qtvlm/code/branches/qtVlm_dev@1985 2387b160-4621-4e8a-be36-40ea10d72da2

+++ src/Grib.h
@@ -62,7 +62,7 @@ class Grib: public QObject
         FCT_GET(QString,fileName)
         FCT_GET(long,fileSize)
 
-        int get_nbDate(void) { return tList.size(); }
+        int get_nbDate(void) { return (int) tList.size(); }
 
         FCT_GET(int,dewpointDataStatus)
         enum GribFileDataStatus { DATA_IN_FILE, NO_DATA_IN_FILE, COMPUTED_DATA };
